调整字段的类型F

+++ space-api/dto/social_media.dto.go
@@ -21,7 +21,7 @@ type (
 
 	CreateOrUpdateSocialMediaReq struct {
 		Id          int64  `json:"id"`
-		Hide        byte   `json:"hide"`
+		Hide        int    `json:"hide"`
 		DisplayName string `gorm:"type:varchar(255);not null;comment:显示名称" json:"displayName"`
 		IconURL     string `gorm:"type:varchar(255);not null;comment:图标链接" json:"iconURL"`
 		OpenUrl     string `gorm:"type:varchar(255);not null;comment:跳转链接" json:"openUrl"`
